chore: refactor out in-line import
