+++ src/macaron/slsa_analyzer/analyzer.py
@@ -43,6 +43,7 @@ from macaron.repo_finder.provenance_extractor import (
 )
 from macaron.repo_finder.provenance_finder import ProvenanceFinder, find_provenance_from_ci
 from macaron.repo_verifier.repo_verifier import verify_repo
+from macaron.repo_finder.repo_finder import get_git_service, prepare_repo
 from macaron.slsa_analyzer import git_url
 from macaron.slsa_analyzer.analyze_context import AnalyzeContext
 from macaron.slsa_analyzer.asset import VirtualReleaseAsset
@@ -54,6 +55,7 @@ from macaron.slsa_analyzer.ci_service import CI_SERVICES
 from macaron.slsa_analyzer.database_store import store_analyze_context_to_db
 from macaron.slsa_analyzer.git_service import GIT_SERVICES, BaseGitService
 from macaron.slsa_analyzer.git_service.base_git_service import NoneGitService
+from macaron.slsa_analyzer.git_url import GIT_REPOS_DIR
 from macaron.slsa_analyzer.package_registry import PACKAGE_REGISTRIES
 from macaron.slsa_analyzer.provenance.expectations.expectation_registry import ExpectationRegistry
 from macaron.slsa_analyzer.provenance.intoto import InTotoPayload, InTotoV01Payload
@@ -68,9 +70,6 @@ logger: logging.Logger = logging.getLogger(__name__)
 
 class Analyzer:
     """This class is used to analyze SLSA levels of a Git repo."""
-
-    GIT_REPOS_DIR = "git_repos"
-    """The directory in the output dir to store all cloned repositories."""
 
     def __init__(self, output_path: str, build_log_path: str) -> None:
         """Initialize instance.
@@ -103,17 +102,6 @@ class Analyzer:
         self.build_log_path = build_log_path
         if not os.path.isdir(self.build_log_path):
             os.makedirs(self.build_log_path)
-
-        # If provided with local_repos_path, we resolve the path of the target repo
-        # to the path within local_repos_path.
-        # If not, we use the default value <output_path>/git_repos/local_repos.
-        self.local_repos_path = (
-            global_config.local_repos_path
-            if global_config.local_repos_path
-            else os.path.join(global_config.output_path, Analyzer.GIT_REPOS_DIR, "local_repos")
-        )
-        if not os.path.exists(self.local_repos_path):
-            os.makedirs(self.local_repos_path, exist_ok=True)
 
         # Load the expectations from global config.
         self.expectations = ExpectationRegistry(global_config.expectation_paths)
@@ -384,8 +372,8 @@ class Analyzer:
         # Prepare the repo.
         git_obj = None
         if analysis_target.repo_path:
-            git_obj = self.prepare_repo(
-                os.path.join(self.output_path, self.GIT_REPOS_DIR),
+            git_obj = prepare_repo(
+                os.path.join(self.output_path, GIT_REPOS_DIR),
                 analysis_target.repo_path,
                 analysis_target.branch,
                 analysis_target.digest,
@@ -859,182 +847,6 @@ class Analyzer:
 
         return analyze_ctx
 
-    def prepare_repo(
-        self,
-        target_dir: str,
-        repo_path: str,
-        branch_name: str = "",
-        digest: str = "",
-        purl: PackageURL | None = None,
-    ) -> Git | None:
-        """Prepare the target repository for analysis.
-
-        If ``repo_path`` is a remote path, the target repo is cloned to ``{target_dir}/{unique_path}``.
-        The ``unique_path`` of a repository will depend on its remote url.
-        For example, if given the ``repo_path`` https://github.com/org/name.git, it will
-        be cloned to ``{target_dir}/github_com/org/name``.
-
-        If ``repo_path`` is a local path, this method will check if ``repo_path`` resolves to a directory inside
-        ``Analyzer.local_repos_path`` and to a valid git repository.
-
-        Parameters
-        ----------
-        target_dir : str
-            The directory where all remote repository will be cloned.
-        repo_path : str
-            The path to the repository, can be either local or remote.
-        branch_name : str
-            The name of the branch we want to checkout.
-        digest : str
-            The hash of the commit that we want to checkout in the branch.
-        purl : PackageURL | None
-            The PURL of the analysis target.
-
-        Returns
-        -------
-        Git | None
-            The pydriller.Git object of the repository or None if error.
-        """
-        # TODO: separate the logic for handling remote and local repos instead of putting them into this method.
-        logger.info(
-            "Preparing the repository for the analysis (path=%s, branch=%s, digest=%s)",
-            repo_path,
-            branch_name,
-            digest,
-        )
-
-        resolved_local_path = ""
-        is_remote = git_url.is_remote_repo(repo_path)
-
-        if is_remote:
-            logger.info("The path to repo %s is a remote path.", repo_path)
-            resolved_remote_path = git_url.get_remote_vcs_url(repo_path)
-            if not resolved_remote_path:
-                logger.error("The provided path to repo %s is not a valid remote path.", repo_path)
-                return None
-
-            git_service = self.get_git_service(resolved_remote_path)
-            repo_unique_path = git_url.get_repo_dir_name(resolved_remote_path)
-            resolved_local_path = os.path.join(target_dir, repo_unique_path)
-            logger.info("Cloning the repository.")
-            try:
-                git_service.clone_repo(resolved_local_path, resolved_remote_path)
-            except CloneError as error:
-                logger.error("Cannot clone %s: %s", resolved_remote_path, str(error))
-                return None
-        else:
-            logger.info("Checking if the path to repo %s is a local path.", repo_path)
-            resolved_local_path = self._resolve_local_path(self.local_repos_path, repo_path)
-
-        if resolved_local_path:
-            try:
-                git_obj = Git(resolved_local_path)
-            except InvalidGitRepositoryError:
-                logger.error("No git repo exists at %s.", resolved_local_path)
-                return None
-        else:
-            logger.error("Error happened while preparing the repo.")
-            return None
-
-        if git_url.is_empty_repo(git_obj):
-            logger.error("The target repository does not have any commit.")
-            return None
-
-        # Find the digest and branch if a version has been specified
-        if not digest and purl and purl.version:
-            found_digest = find_commit(git_obj, purl)
-            if not found_digest:
-                logger.error(
-                    "Could not map the input purl string to a specific commit in the corresponding repository."
-                )
-                return None
-            digest = found_digest
-
-        # Checking out the specific branch or commit. This operation varies depends on the git service that the
-        # repository uses.
-        if not is_remote:
-            # If the repo path provided by the user is a local path, we need to get the actual origin remote URL of
-            # the repo to decide on the suitable git service.
-            origin_remote_url = git_url.get_remote_origin_of_local_repo(git_obj)
-            if git_url.is_remote_repo(origin_remote_url):
-                # The local repo's origin remote url is a remote URL (e.g https://host.com/a/b): In this case, we obtain
-                # the corresponding git service using ``self.get_git_service``.
-                git_service = self.get_git_service(origin_remote_url)
-            else:
-                # The local repo's origin remote url is a local path (e.g /path/to/local/...). This happens when the
-                # target repository is a clone from another local repo or is a clone from a git archive -
-                # https://git-scm.com/docs/git-archive: In this case, we fall-back to the generic function
-                # ``git_url.check_out_repo_target``.
-                if not git_url.check_out_repo_target(git_obj, branch_name, digest, not is_remote):
-                    logger.error("Cannot checkout the specific branch or commit of the target repo.")
-                    return None
-
-                return git_obj
-
-        try:
-            git_service.check_out_repo(git_obj, branch_name, digest, not is_remote)
-        except RepoCheckOutError as error:
-            logger.error("Failed to check out repository at %s", resolved_local_path)
-            logger.error(error)
-            return None
-
-        return git_obj
-
-    @staticmethod
-    def get_git_service(remote_path: str | None) -> BaseGitService:
-        """Return the git service used from the remote path.
-
-        Parameters
-        ----------
-        remote_path : str | None
-            The remote path of the repo.
-
-        Returns
-        -------
-        BaseGitService
-            The git service derived from the remote path.
-        """
-        if remote_path:
-            for git_service in GIT_SERVICES:
-                if git_service.is_detected(remote_path):
-                    return git_service
-
-        return NoneGitService()
-
-    @staticmethod
-    def _resolve_local_path(start_dir: str, local_path: str) -> str:
-        """Resolve the local path and check if it's within a directory.
-
-        This method returns an empty string if there are errors with resolving ``local_path``
-        (e.g. non-existed dir, broken symlinks, etc.) or ``start_dir`` does not exist.
-
-        Parameters
-        ----------
-        start_dir : str
-            The directory to look for the existence of path.
-        local_path: str
-            The local path to resolve within start_dir.
-
-        Returns
-        -------
-        str
-            The resolved path in canonical form or an empty string if errors.
-        """
-        # Resolve the path by joining dir and path.
-        # Because strict mode is enabled, if a path doesn't exist or a symlink loop
-        # is encountered, OSError is raised.
-        # ValueError is raised if we use both relative and absolute paths in os.path.commonpath.
-        try:
-            dir_real = os.path.realpath(start_dir, strict=True)
-            resolve_path = os.path.realpath(os.path.join(start_dir, local_path), strict=True)
-            if os.path.commonpath([resolve_path, dir_real]) != dir_real:
-                return ""
-
-            return resolve_path
-        except (OSError, ValueError) as error:
-            logger.error(error)
-            return ""
-
     def _determine_git_service(self, analyze_ctx: AnalyzeContext) -> BaseGitService:
         """Determine the Git service used by the software component."""
         remote_path = analyze_ctx.component.repository.remote_path if analyze_ctx.component.repository else None
@@ -1096,8 +908,8 @@ class Analyzer:
             ci_service.set_api_client()
 
             if ci_service.is_detected(
-                repo_path=analyze_ctx.component.repository.fs_path,
-                git_service=analyze_ctx.dynamic_data["git_service"],
+                    repo_path=analyze_ctx.component.repository.fs_path,
+                    git_service=analyze_ctx.dynamic_data["git_service"],
             ):
                 logger.info("The repo uses %s CI service.", ci_service.name)
 
@@ -1125,7 +937,7 @@ class Analyzer:
     def _determine_package_registries(self, analyze_ctx: AnalyzeContext) -> None:
         """Determine the package registries used by the software component based on its build tools."""
         build_tools = (
-            analyze_ctx.dynamic_data["build_spec"]["tools"] or analyze_ctx.dynamic_data["build_spec"]["purl_tools"]
+                analyze_ctx.dynamic_data["build_spec"]["tools"] or analyze_ctx.dynamic_data["build_spec"]["purl_tools"]
         )
         for package_registry in PACKAGE_REGISTRIES:
             for build_tool in build_tools:
