+++ tests/slsa_analyzer/test_analyzer.py
@@ -3,8 +3,6 @@
 
 """This module tests the slsa_analyzer.Gh module."""
 
-from pathlib import Path
-
 import hypothesis.provisional as st_pr
 import hypothesis.strategies as st
 import pytest
@@ -14,35 +12,6 @@ from packageurl import PackageURL
 from macaron.config.target_config import Configuration
 from macaron.errors import InvalidAnalysisTargetError, InvalidPURLError
 from macaron.slsa_analyzer.analyzer import Analyzer
-
-from ..macaron_testcase import MacaronTestCase
-
-
-class TestAnalyzer(MacaronTestCase):
-    """
-    This class contains all the tests for the Analyzer
-    """
-
-    # Using the parent dir of this module as a valid start dir.
-    PARENT_DIR = str(Path(__file__).parent)
-
-    # pylint: disable=protected-access
-    def test_resolve_local_path(self) -> None:
-        """Test the resolve local path method."""
-        # Test resolving a path outside of the start_dir
-        assert not Analyzer._resolve_local_path(self.PARENT_DIR, "../")
-        assert not Analyzer._resolve_local_path(self.PARENT_DIR, "./../")
-        assert not Analyzer._resolve_local_path(self.PARENT_DIR, "../../../../../")
-
-        # Test resolving a non-existing dir
-        assert not Analyzer._resolve_local_path(self.PARENT_DIR, "./this-should-not-exist")
-
-        # Test with invalid start_dir
-        assert not Analyzer._resolve_local_path("non-existing-dir", "./")
-
-        # Test resolve successfully
-        assert Analyzer._resolve_local_path(self.PARENT_DIR, "./") == self.PARENT_DIR
-        assert Analyzer._resolve_local_path(self.PARENT_DIR, "././././") == self.PARENT_DIR
 
 
 @pytest.mark.parametrize(
